Bundle nvcuda.dll
This is a pretty dumb issue:

On Windows, CUDA is dynamically linked via nvcuda.dll.
However, CI runners don't have this DLL as it's part of the NVIDIA GPU driver.
This DLL is not needed for people with NVIDIA GPUs either as they already have it in System32.
So it's only needed to avoid undefined symbols for non-CUDA GPUs, not to actually use it.

This is why we have to bundle it ourselves.